Replace controller files by ControllerProvier classes
